created API endpoint for checking validity of JWT tokens. added useEffect to redirect unauthorized users
