+++ client/package.json
@@ -6,9 +6,9 @@
   "scripts": {
     "dev": "vite",
     "build": "vite build",
+    "preview": "vite preview",
     "lint": "biome lint ./src",
-    "fmt": "biome format --write ./src",
-    "preview": "vite preview"
+    "fmt": "biome format --write ./src"
   },
   "dependencies": {
     "@headlessui/react": "^2.1.9",
