+++ client/src/components/button.jsx
@@ -1,0 +1,3 @@
+export default function Button({ children, ...props }) {
+  return <button {...props}>{children}</button>;
+}
