--- client/src/majorDetails
@@ -1,39 +1,0 @@
-import React from 'react';
-import { useParams, useNavigate } from 'react-router-dom';
-import './webPage.css';
-
-const MajorDetails = () => {
-    const { majorId } = useParams(); // Get the majorId from the URL params
-    const navigate = useNavigate();
-
-    // Placeholder data. Replace this with actual data from your state or API.
-    const majorsData = [
-        { id: '1', name: 'Computer Science', description: 'Study of computation and information systems.' },
-        { id: '2', name: 'Business Administration', description: 'Management of business operations and activities.' },
-        { id: '3', name: 'Biology', description: 'The study of living organisms.' },
-    ];
-
-    // Find the major by ID
-    const major = majorsData.find((m) => m.id === majorId);
-
-    const handleBack = () => {
-        navigate('/admin-dash'); // Navigate back to the Admin Dashboard
-    };
-
-    return (
-        <div className="major-details">
-            <h1>Major Details</h1>
-            {major ? (
-                <div>
-                    <h2>{major.name}</h2>
-                    <p>{major.description}</p>
-                </div>
-            ) : (
-                <p>No major found with the given ID.</p>
-            )}
-            <button onClick={handleBack}>Back to Admin Dashboard</button>
-        </div>
-    );
-};
-
-export default MajorDetails;
