+++ client/src/pages/teacher/TeacherDashboard.jsx
@@ -5,11 +5,12 @@ import {
   DialogPanel,
   DialogTitle,
 } from "@headlessui/react";
-import { useState } from "react";
+import { useEffect, useState } from "react";
 import { useNavigate } from "react-router-dom";
-import "../stylesheets/webPage.css";
-import { JWT_KEY, useLocalState } from "../hooks/useLocalStorage";
-import { apiFetch } from "../utils/apiFetch";
+import "../../styles/webPage.css";
+import { JWT_KEY, useLocalState } from "../../hooks/useLocalStorage";
+import { apiChangePassword, apiCheckToken } from "../../lib/api";
+import { decodeJWT } from "../../utils/decodeJWT";
 
 const TeacherDash = () => {
   const [jwt, setJwt] = useLocalState("", JWT_KEY);
@@ -21,40 +22,57 @@ const TeacherDash = () => {
 
   const handleLogout = () => {
     localStorage.removeItem(JWT_KEY);
-
     navigate("/");
   };
 
-  const handleChangePassword = async (e) => {
-    e.preventDefault();
+  const handleChangePassword = async (event) => {
+    event.preventDefault();
 
     if (currentPassword === newPassword) {
-        alert("passwords must be different");
-        return;
+      alert("passwords must be different");
+      return;
     }
 
-    const params = {
-        url: "http://localhost:8000/v1/auth/change-password",
-        method: "POST",
-        jwt,
-        requestBody: {
-            current_password: currentPassword,
-            new_password: newPassword,
-        },
-    };
-
-    const response = await apiFetch(params);
+    const response = await apiChangePassword(currentPassword, newPassword, jwt);
 
     if (response.error) {
-        alert(`${response.error}`);
+      alert(`${response.error}`);
     } else {
-        alert(`${response.success}`);
+      alert(`${response.success}`);
 
-        setCurrentPassword("");
-        setNewPassword("");
-        setIsOpen(false);
+      setCurrentPassword("");
+      setNewPassword("");
+      setIsOpen(false);
     }
   };
+
+  useEffect(() => {
+    const checkToken = async () => {
+      if (jwt) {
+        const response = await apiCheckToken(jwt);
+
+        if (response.error) {
+          navigate("/");
+        }
+
+        const decodedRole = decodeJWT(jwt).role;
+
+        if (decodedRole !== "INSTRUCTOR") {
+          navigate(
+            decodedRole === "ADMIN"
+              ? "/admin"
+              : decodedRole === "STUDENT"
+                ? "/student"
+                : "/",
+          );
+        }
+      } else {
+        navigate("/");
+      }
+    };
+
+    checkToken();
+  }, [jwt, navigate]);
 
   return (
     <div className="admin-dashboard">
@@ -115,12 +133,13 @@ const TeacherDash = () => {
                       className="form-button"
                       type="button"
                       onClick={() => setIsOpen(false)}
-                    >Cancel</button>
+                    >
+                      Cancel
+                    </button>
 
                     <button className="form-button" type="submit">
                       Confirm
                     </button>
-
                   </div>
                 </form>
               </DialogPanel>
