+++ server/src/api/routes/authRoutes.js
@@ -1,8 +1,8 @@
-import { changePassword, login } from "../handlers/authHandler.js";
+import { changePassword, checkToken, login } from "../handlers/authHandler.js";
 
 function authRoutes(server) {
+  server.get("/v1/auth/check", async (ctx) => await checkToken(ctx));
   server.post("/v1/auth/login", async (ctx) => await login(ctx));
-
   server.post(
     "/v1/auth/change-password",
     async (ctx) => await changePassword(ctx),
